Fixes to generator properties and init_solver().
Added property type hints to the abstract `Generator` class.  Stumbled
on these when checking some calling code with mypy.

This caused me to fix the `init_solver` method in gen2.py.
Previously, this would only set the solver property if
`settings.single_solve` was true.  But gen2 is only used if
`single_solve` is true, so I removed that condition.  If that
condition was false, then `init_solver` would fail in odd ways because
it would set properties on the `solver` but `solver` would be unbound.