Fix sorting of Popper-generated rules.
Previously, these were always sorted by `format_prog` and
`format_rule`.  Now this is controlled by an argument that defaults
to *not* sorting.  We check for the use of recursion or direction in
the bias, and if these are present, we do not sort, which avoids
problems in the tester.